feat(file-transfer): Allow disabling of fitbit-ga4 auto file transfer processing
As soon as a file transfer is popped on the companion's side, that file
is handled and wont be visible to other companion. This is problematic
if a module such as fitbit-ga4 processes files first, or if the main
companion processes first. One of them will not receive its expected
files. If fitbit-ga4 doesn't get its files, you risk losing data.

This code change allows to disable fitbit-ga4 auto processing and
require the main companion to manually pass us the files they do not
recognize instead.